Move things from `conftest` to `tests/auxil`
