+++ tests/auxil/bot_method_checks.py
@@ -38,7 +38,7 @@ from telegram._utils.defaultvalue import DEFAULT_NONE, DefaultValue
 from telegram.constants import InputMediaType
 from telegram.ext import Defaults, ExtBot
 from telegram.request import RequestData
-from tests.auxil.object_conversions import env_var_2_bool
+from tests.auxil.envvars import env_var_2_bool
 
 TEST_WITH_OPT_DEPS = env_var_2_bool(os.getenv("TEST_WITH_OPT_DEPS", True))
 if TEST_WITH_OPT_DEPS:
@@ -338,9 +338,14 @@ async def check_defaults_handling(
                         pytest.fail(f"ILQR.i_m_c has a {attr}, expected it to be absent")
                 # Here we explicitly use that we only pass InputTextMessageContent for testing
                 # which has both attributes
-                elif imc.get(attr) != df_value:
+                elif (
+                    imc.get(
+                        attr,
+                    )
+                    != df_value
+                ):
                     pytest.fail(
-                        f"Got value {imc.get(attr)} for ILQR.i_m_c.{attr} instead of {df_value}"
+                        f"Got value {imc.get(attr, )} for ILQR.i_m_c.{attr} instead of {df_value}"
                     )
 
         # Check datetime conversion
