+++ tests/auxil/ci_bots.py
@@ -22,6 +22,11 @@ import json
 import os
 import random
 
+from tests.auxil.constants import PRIVATE_KEY
+from tests.auxil.envvars import TEST_WITH_OPT_DEPS
+from tests.auxil.networking import NonchalantHttpxRequest
+from tests.auxil.pytest_classes import PytestExtBot
+
 # Provide some public fallbacks so it's easy for contributors to run tests on their local machine
 # These bots are only able to talk in our test chats, so they are quite useless for other
 # purposes than testing.
@@ -52,20 +57,37 @@ class BotInfoProvider:
     def __init__(self):
         self._cached = {}
 
+    @staticmethod
+    def _get_value(key, fallback):
+        # If we're running as a github action then fetch bots from the repo secrets
+        if GITHUB_ACTION is not None and BOTS is not None and JOB_INDEX is not None:
+            try:
+                return BOTS[JOB_INDEX][key]
+            except (IndexError, KeyError):
+                pass
+
+        # Otherwise go with the fallback
+        return fallback
+
     def get_info(self):
         if self._cached:
             return self._cached
-        self._cached = {k: get(k, v) for k, v in random.choice(FALLBACKS).items()}
+        self._cached = {k: self._get_value(k, v) for k, v in random.choice(FALLBACKS).items()}
         return self._cached
 
 
-def get(key, fallback):
-    # If we're running as a github action then fetch bots from the repo secrets
-    if GITHUB_ACTION is not None and BOTS is not None and JOB_INDEX is not None:
-        try:
-            return BOTS[JOB_INDEX][key]
-        except (IndexError, KeyError):
-            pass
-
-    # Otherwise go with the fallback
-    return fallback
+def make_bot(bot_info=None, **kwargs):
+    """
+    Tests are executed on tg.ext.ExtBot, as that class only extends the functionality of tg.bot
+    """
+    token = kwargs.pop("token", (bot_info or {}).get("token"))
+    private_key = kwargs.pop("private_key", PRIVATE_KEY)
+    kwargs.pop("token", None)
+    _bot = PytestExtBot(
+        token=token,
+        private_key=private_key if TEST_WITH_OPT_DEPS else None,
+        request=NonchalantHttpxRequest(8),
+        get_updates_request=NonchalantHttpxRequest(1),
+        **kwargs,
+    )
+    return _bot
