+++ tests/auxil/plugin_github_group.py
