+++ tests/test_meta.py
@@ -20,7 +20,7 @@ import os
 
 import pytest
 
-from tests.auxil.object_conversions import env_var_2_bool
+from tests.auxil.envvars import env_var_2_bool
 
 skip_disabled = pytest.mark.skipif(
     not env_var_2_bool(os.getenv("TEST_BUILD", False)), reason="TEST_BUILD not enabled"
