+++ tests/test_requestparameter.py
@@ -24,7 +24,7 @@ import pytest
 from telegram import InputFile, InputMediaPhoto, InputMediaVideo, MessageEntity
 from telegram.constants import ChatType
 from telegram.request._requestparameter import RequestParameter
-from tests.conftest import data_file
+from tests.auxil.files import data_file
 
 
 class TestRequestParameterWithoutRequest:
